Fix crash in FLAnimatedImageView when stopAnimating is called from -[UIImageView dealloc]
-startAnimating is the only place where we need to lazily instantiate the displayLink, so we’ve moved the logic there. In the other places where we call the displayLink getter (i.e. -stopAnimating and -isAnimating), we don’t want to create a display link if we don’t have one already.

Creating the displayLink and the weak proxy when the image view has already entered deallocation is problematic because the weak system won’t allow forming new weak references to the deallocating object. It’s also unnecessary work that we don’t want to be doing while the object is going away.