Check for both the definition and value of DEBUG in FLAnimatedImage test project classes.
Vanilla iOS projects define DEBUG to be 1 in the Debug build configuration, and don't define DEBUG at all in the Release build configuration. That means that if you make a release build DEBUG is implicitly 0, but if you turn on the -Wundef warning flag you'll get warnings about this implicit value. This change explicitly makes sure both that DEBUG is defined and that its value evaluates to true.