Refactor FLAnimatedImage to support pluggable image modules.
A number of refactors to make FLAnimatedImage more supporting of new
image formats. First, split the caching logic into its own object,
FLAnimatedImageFrameCache, so it can be shared between types. Next,
introduce an abstraction over the imate data, FLAnimatedImageData.
Finally, update the GIF support to be implemented as a plugin for
the core FLAnimatedImage library.