Remove debug_info property from FLAnimatedImage.h.
This was only included in DEBUG builds, but when FLAnimatedImage was built as a framework then used in a project that was being built as DEBUG is would falsely show up as available and cause crashes at runtime. This was used within the Flipboard app, but not elsewhere in practice, and can be replaced with categories in specific apps if desired.